convert Listing into container and refactor ListItem into component
