+++ client/components/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import List from './List'
+import List from '../containers/List'
 
 function App (props) {
   return (
